requirements.txt
python-telegram-bot
flask
asyncio
threading